GeoMesa Project
Requires a version of the GeoMesa client library that has been
backported to Scala 2.10.  This is to make it easier to run the
client (synthetic data generator) on EMR.

In order to compile this code, you must have GeoMesa 1.2.4-BACKPORT in
your local Maven repository.  To do that, pull and publish-m2 [1] then
pull and 'mvn install' [2].

[1] https://github.com/jamesmcclain/sfcurve/tree/feature/scala_2.10.6
[2] https://github.com/jamesmcclain/geomesa/tree/feature/scala_2.10.6